Frameworks: add all slim shortcut image resources
Importing for now the old one from KitKat. Graphics group is currently
vectorizing and revamping all. When they are finished we overwrite them and
add xxxhdpi ones as well.

Doing this now due that it is stopping the progress on a bunch of features right now

And a small method renaming to be inline with the other methods

Change-Id: If07e2ae19a45712cd537389069d912cf425e9156

+

Add missing file

core/java/com/android/internal/util/crdroid/ActionHelper.java